Do not set ECID as primary identity. (#21)
* Do not set ECID as primary identity.
Do not automatically set any identity as primary. This change sets the 'primary' ECID as 'primary = false'. This has the side effect of not allowing any secondary ECID if the primary is not set.  The reason for this change is there can only be one primary identity in an identity map. If a custom identity is added which is set as primary, then the ECID cannot be primary. The Experience Edge Network (Konductor) will set the first ECID in the identity map as primary if no other identity is marked as primary. In this case, the client implementation is defering to the server-side to determine the primary identity in the request.

* Update CocoaPods to get latest from Edge

* Simplify adding primary ECID by adding function in IdentityMap to insert an item to the front of the list.